Fix CountSetsReps feature
