--- data/workoutRecording.txt
@@ -1,18 +1,0 @@
-Date    : 02/02/21
-Workout : run2
-Exercise: 
-          1. Name: running, Weight: 45kg, Reps: 4 4 4 4
-          2. Name: running, Weight: 45kg, Reps: 4 4 4 4
-Workout : run
-Exercise: 
-          1. Name: running, Weight: 45kg, Reps: 4 4 4 4
-          2. Name: running, Weight: 45kg, Reps: 4 4 4 4
-          3. Name: running, Weight: 45kg, Reps: 4 4 4 4
-          4. Name: running, Weight: 45kg, Reps: 4 4 4 4
-Date    : 03/03/21
-Workout : w 3232 332
-Exercise: 
-Workout : www
-Exercise: 
-Workout : w 3
-Exercise:
